split off useOverflow hook
